ARPA-216 Moved swagger classes to separate folder
