+++ Orso.Arpa.Api/Swagger/SwaggerAuthorizeOperationFilter.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
-namespace Orso.Arpa.Api.ModelBinding
+namespace Orso.Arpa.Api.Swagger
 {
     public class SwaggerAuthorizeOperationFilter : IOperationFilter
     {
